requirements.txt
# Emotion-Aware Broadcast News Analytics — dependencies
# Python 3.10+ recommended

# Data and numerics
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Visualization (for analysis/plotting scripts)
matplotlib>=3.7.0

# NLP and topic modeling
transformers>=4.30.0
torch>=2.0.0
gensim>=4.3.0
nltk>=3.8.0
scipy>=1.10.0
spacy>=3.5.0

# YouTube and video
google-api-python-client>=2.100.0
pytubefix>=2.0.0
yt-dlp>=2023.0.0
moviepy>=1.0.3

# Transcription and segmentation
assemblyai>=0.17.0
langchain-ai21>=0.0.1

# Rephrasing / generation
google-generativeai>=0.3.0

# Image (if using image.py)
Pillow>=9.5.0
httpx>=0.24.0

# Download spaCy model after install:
# python -m spacy download en_core_web_sm
